seeder/pyproject.toml
[project]
name = "seeder"
version = "0.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Damy Metzke", email = "damy.metzke@gmail.com"},
]
dependencies = [
    "psycopg2>=2.9.9",
    "Faker>=25.8.0",
    "argon2-cffi>=23.1.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}


# [tool.pdm]
# distribution = false
